move shape indicator overay and improve example
