+++ apps/examples/src/examples/ui/indicators-logic/README.md
@@ -15,8 +15,10 @@ keywords:
   ]
 ---
 
-Change when indicators are shown and how they appear.
+Change when shape indicators are shown and how they appear.
 
 ---
 
-This example shows how you can change when indicators are shown and how they appear.
+This example shows how you can change when shape indicators are shown and how they appear.
+
+Shape indicators are the lines that are drawn around the border of a shape when hovering over it. This examples makes them appear all of the time.
